% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_prealloc.R
\name{serialize_prealloc}
\alias{serialize_prealloc}
\title{Serialize with preallocated buffer.}
\usage{
serialize_prealloc(object, prealloc = TRUE, trace = FALSE)
}
\arguments{
\item{object}{Object for serialization.}

\item{prealloc}{Should preallocation be used?}

\item{trace}{Print traces for buffer memory allocations.}
}
\description{
Serialize with preallocated buffer.
}
